requirements.txt
Flask==2.3.2
ultralytics>=8.3
python-dotenv
